Modify player config
